+++ pybmd/__init__.py
@@ -5,7 +5,7 @@ __name__ = 'pybmd'
 
 __major_version__ = "2024"
 __minor_version__ = "3"
-__change_version__ = "0"
+__change_version__ = "1"
 __version__ = ".".join(
     (__major_version__, __minor_version__, __change_version__)
 )